+++ src/main/java/subway/line/application/LineService.java
@@ -1,10 +1,14 @@
-package subway.line;
+package subway.line.application;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import subway.line.Line;
+import subway.line.infra.LineRepository;
+import subway.line.controller.LineRequest;
+import subway.line.controller.LineResponse;
 import subway.line.exception.LineNotFoundException;
 import subway.station.Station;
-import subway.station.StationService;
+import subway.station.application.StationService;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -34,7 +38,15 @@ public class LineService {
                         downStation.getId(),
                         lineRequest.getDistance()));
 
-        return new LineResponse(line);
+        /*
+        * presentation 에서 외부로 노출되는 클래스는 가장 변화가 일어나기 쉽기 때문에, use case 마다 분리하는 것이 좋은 경우가 많습니다.
+        * use case 마다 분리
+        * */
+        return new LineResponse(
+                line.getId(),
+                line.getName(),
+                line.getColor(),
+                findStations(line.getUpStationId(), line.getDownStationId()));
     }
 
     public List<LineResponse> findAllLines() {
