+++ src/main/java/subway/line/controller/LineRequest.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.line.controller;
 
 public class LineRequest {
