+++ src/main/java/subway/station/application/StationService.java
@@ -1,8 +1,12 @@
-package subway.station;
+package subway.station.application;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import subway.station.Station;
+import subway.station.controller.StationRequest;
+import subway.station.controller.StationResponse;
 import subway.station.exception.StationNotFoundException;
+import subway.station.infra.StationRepository;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -36,6 +40,10 @@ public class StationService {
     }
 
     public StationResponse findStationById(Long id) {
+        /*
+        * StationResponse는 응용 계층에게 전달하기 위한 Dto인데 도메인 간에 사용하는 Dto를 따로 만드는 편이 좋을 것 같다.
+        * 네이밍을 어떻게 해야하지..?
+        * */
         Station station = stationRepository.findById(id).orElseThrow(StationNotFoundException::new);
         return createStationResponse(station);
     }
