+++ src/main/java/subway/station/controller/StationController.java
@@ -1,7 +1,8 @@
-package subway.station;
+package subway.station.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import subway.station.application.StationService;
 
 import java.net.URI;
 import java.util.List;
