+++ src/main/java/subway/station/controller/StationRequest.java
@@ -1,4 +1,4 @@
-package subway.station;
+package subway.station.controller;
 
 public class StationRequest {
     private String name;
