+++ src/main/java/subway/station/infra/StationRepository.java
@@ -1,6 +1,7 @@
-package subway.station;
+package subway.station.infra;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import subway.station.Station;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
 }
